% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_complex_header_above}
\alias{add_complex_header_above}
\title{Add complex header above the kable table}
\usage{
add_complex_header_above(kable_input, tbl, row.names = NA)
}
\arguments{
\item{kable_input}{is whatever kable input.}

\item{tbl}{is the initial table.}

\item{row.names}{shoul we include `row.names`?}
}
\description{
Add complex header above the kable table. It is supposed to be a part of `knitr - kable - kableExtra` pipeline. Relies on `kableExtra::add_header_above` when `behead` returns a table with complex header.
(E.g., when the table is `mlth.data.frame`.)
}
