% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{behead}
\alias{behead}
\title{Separate table header and table body}
\usage{
behead(tbl)
}
\arguments{
\item{tbl}{is a `mlth.data.frame` or `data.frame`. If `tbl` is a `data.frame`, the function returns it unchanged.}
}
\description{
Separate table header and table body to write it to a spreadsheet or to html or to whatever.
It returns dataframe with a header as an `attr(tbl, 'header')``.
}
\details{
Also see `unpivotr::behead()``
}
