% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{is.mlth.data.frame}
\alias{is.mlth.data.frame}
\alias{as.mlth.data.frame}
\alias{as.mlth.data.frame.default}
\alias{as.mlth.data.frame.list}
\alias{as.mlth.data.frame.matrix}
\alias{as.list.mlth.data.frame}
\title{Coerce to a Multi-Header Data Frame}
\usage{
is.mlth.data.frame(x)

as.mlth.data.frame(x, row.names = NULL, ...)

\method{as.mlth.data.frame}{default}(x, ...)

\method{as.mlth.data.frame}{list}(x, ...)

\method{as.mlth.data.frame}{matrix}(x, ...)

\method{as.list}{mlth.data.frame}(x)
}
\arguments{
\item{x}{any \R{} object}

\item{row.names}{\code{NULL} or character or integer vector to be used as row names, must be of the same
length as the variables.}

\item{...}{other arguments passed \code{mlth.data.frame}.}
}
\value{
\code{is.mlth.data.frame} returns a multi-header data frame.
\code{as.mlth.data.frame} returns \code{TRUE} if the object has
"\code{mlth.data.frame}" within its classes and \code{FALSE} otherwise.
}
\description{
Functions to check and coerce the object to a \code{mlth.data.frame}.
}
\examples{
L <- list(
		X = c('A', 'B', 'C'),
		Y = list(
			       N = 1:3,
			       M = 4:6))

(A <- as.mlth.data.frame(L))
is.mlth.data.frame(A)

}
