% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{kable_cors}
\alias{kable_cors}
\alias{mlth_cors}
\title{Render correlation table}
\usage{
kable_cors(x, y = x, type = c("pearson", "spearman"))

mlth_cors(x, y = x, type = c("pearson", "spearman"))
}
\arguments{
\item{x, y}{are tables (`matrix`, `data.frame`).}

\item{type}{is type of correlation: Pearson or Spearman.}
}
\description{
Render correlation table either as `mlth.data.frame` or as `kable` table.
}
\details{
When using `kable_cors`, include the following html-code to turn on popovers:
`<!--html_preserve-->`
`<script>`
`$(document).ready(function(){`
`   $('[data-toggle="popover"]').popover();` 
`});`
`</script>`
`<!--/html_preserve-->`
}
