% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{mlth.data.frame}
\alias{mlth.data.frame}
\alias{mlth}
\title{Multi-Header Data Frames}
\usage{
mlth.data.frame(
  ...,
  row.names = NULL,
  check.rows = FALSE,
  check.names = FALSE,
  fix.empty.names = FALSE
)

mlth(
  ...,
  row.names = NULL,
  check.rows = FALSE,
  check.names = FALSE,
  fix.empty.names = FALSE
)
}
\arguments{
\item{...}{the collection of named variables or sub-tables, each of class \code{list}. 
All variables must be of the same length.}

\item{row.names}{\code{NULL} or character or integer vector to be used as row names, must be of the same
length as the variables.}

\item{stringsAsFactors}{same as for \code{\link{data.frame}}}

\item{defaultName}{the single character value to fill empty names.}

\item{fixNamesSep}{the separator which is used when there are more than one empty names.}
}
\value{
\code{mlth.data.frame} object
}
\description{
Create a multi-header data frame which is 
an hierarchical collection of variables.
}
\details{
A \code{mlth.data.frame} object is a list of variables and \code{mlth.data.frame}s.
Each variable must be a vector of atomic data type (e.g., POSIX date will not work).
Each variable or sub-table must have a name, the names must be unique within
the table/sub-table, but not across sub-tables.

Unlike \code{data.frame}, a multi-header data frame cannot have zero columns and non-zero rows
but can have zero rows and non-zero columns or zero columns and rows.
}
\examples{
A<-mlth.data.frame(X=c('A','B','C'),
			Y=list(
				N=1:3,
				M=4:6))

## The empty names are filled-in
B<-mlth.data.frame(X=list(rnorm(10),rnorm(10)),
			Y=list(rnorm(10),rnorm(10)),
			row.names=letters[1:10])

## str method for mlth.data.frame
str(B)

}
