% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{kable_collapse_rows}
\alias{kable_collapse_rows}
\title{Render a table with layered rows}
\usage{
kable_collapse_rows(l, ...)
}
\arguments{
\item{l}{is a list of tables.}

\item{...}{are parameters passed to kable.}
}
\description{
Render a table with layered rows using kable. 
It is supposed to be a list of tables that define the pieces of the output table.
}
