% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mlth.data.frame}
\alias{print}
\alias{print.mlth.data.frame}
\title{Print Multi-Headed Data Frame}
\usage{
\method{print}{mlth.data.frame}(x, ...)
}
\arguments{
\item{x}{a \code{mlth.data.frame} object.}

\item{...}{other arguments: \code{digits} - number of digits after decimal point 
(default NULL, same as for \code{\link{print.data.frame}}); 
\code{justify} - justify header left, right or centre 
(default 'left', same as 'none').}
}
\description{
The method to print a \code{mlth.data.frame} object.
}

