% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{register_output}
\alias{register_output}
\title{Register table for the output}
\usage{
register_output(tbl, name = NULL, caption = NULL, note = NULL)
}
\arguments{
\item{tbl}{is a `data.frame` or `mlth.data.frame` or any other input supported by `\link{write.xlsx.output}`.}

\item{name}{is table name in the `OUTPUT` list. Can be empty.}

\item{caption}{is table caption as a merged cell above the table.}

\item{note}{is table footnote as a merged cell below the table.}
}
\value{
`tbl` with 'caption' and 'note' attributes
}
\description{
Save the table into a global `OUTPUT` list to write is as an output spreadsheet later.
}
