% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{write.xlsx.output}
\alias{write.xlsx.output}
\title{Write registered output tables}
\usage{
write.xlsx.output(
  file,
  x = list(),
  overwrite = TRUE,
  writer = xlsx.writer.openxlsx
)
}
\arguments{
\item{file}{is the name of `xlsx` file.}

\item{x}{a list of output tables.}

\item{overwrite}{should we overwrite the existing output file?}

\item{writer}{is the function that writes list of tables into an xlsx file.}
}
\description{
Write the contents of `OUTPUT` list to an `xlsx` file. This function is supposed to be used 
at the very end of the analysis when all output tables are prepared.
}
