% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xlsx.writer.openxlsx}
\alias{xlsx.writer.openxlsx}
\title{Write tables to xlsx file}
\usage{
xlsx.writer.openxlsx(tblList, file, overwrite)
}
\arguments{
\item{tblList}{is a list of `data.frame`s. It is assumed that the input table can have `caption` and `note` attributes
and may accept beheaded `mlth.data.frame` (attribute `header`).}

\item{file}{is the name of xlsx file.}

\item{overwrite}{should we overwrite the file?}
}
\description{
These are the writers to use for writing the tables to an xlsx file.
Different writers can rely on different packages, like `openxlsx` or `xlsx`.
My current package of choice is `openxlsx`.
}
\details{
It is important that tblList is a true list! `data.frame` is also a list and
the function will throw an error if `tblList` is `data.frame`.
}
