% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.xlsx}
\alias{write.xlsx}
\alias{write.xlsx.default}
\title{Generic Function \code{write.xlsx}}
\usage{
write.xlsx(x, ...)

\method{write.xlsx}{default}(x, ...)
}
\description{
This package defines the generic function \code{write.xlsx}.
\code{write.xlsx.default} is \code{\link[xlsx]{write.xlsx}}.
}

